Added EGL for pretty printing SQL: found one bug in stage 8
