--- solutions/Epsilon/src/main/resources/metamodels/sql.ecore
@@ -1,93 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ecore:EPackage xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns:ecore="http://www.eclipse.org/emf/2002/Ecore" name="sql" nsURI="http://www.example.org/sql" nsPrefix="sql">
-  <eClassifiers xsi:type="ecore:EClass" name="SelectStatement">
-    <eStructuralFeatures xsi:type="ecore:EReference" name="selectBody" lowerBound="1"
-        eType="#//PlainSelect" containment="true"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="PlainSelect">
-    <eStructuralFeatures xsi:type="ecore:EReference" name="selItems" lowerBound="1"
-        upperBound="-1" eType="#//SelectItem" containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="fromItem" eType="#//FromItem"
-        containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="joins" upperBound="-1"
-        eType="#//Join" containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="whereExp" eType="#//Expression"
-        containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="groupBy" eType="#//GroupByElement"
-        containment="true"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="GroupByElement">
-    <eStructuralFeatures xsi:type="ecore:EReference" name="groupByExps" upperBound="-1"
-        eType="#//Expression" containment="true"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="Join">
-    <eStructuralFeatures xsi:type="ecore:EAttribute" name="left" eType="ecore:EDataType http://www.eclipse.org/emf/2002/Ecore#//EBoolean"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="rightItem" lowerBound="1"
-        eType="#//FromItem" containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="onExp" eType="#//Expression"
-        containment="true"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="Alias">
-    <eStructuralFeatures xsi:type="ecore:EAttribute" name="name" lowerBound="1" eType="ecore:EDataType http://www.eclipse.org/emf/2002/Ecore#//EString"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="Expression" abstract="true" interface="true"/>
-  <eClassifiers xsi:type="ecore:EClass" name="BinaryExpression" abstract="true" eSuperTypes="#//Expression">
-    <eStructuralFeatures xsi:type="ecore:EReference" name="leftExp" lowerBound="1"
-        eType="#//Expression" containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="rightExp" lowerBound="1"
-        eType="#//Expression" containment="true"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="AndExpression" eSuperTypes="#//BinaryExpression"/>
-  <eClassifiers xsi:type="ecore:EClass" name="OrExpression" eSuperTypes="#//BinaryExpression"/>
-  <eClassifiers xsi:type="ecore:EClass" name="EqualsToExpression" eSuperTypes="#//BinaryExpression"/>
-  <eClassifiers xsi:type="ecore:EClass" name="GreaterThanExpression" eSuperTypes="#//BinaryExpression"/>
-  <eClassifiers xsi:type="ecore:EClass" name="SelectItem">
-    <eStructuralFeatures xsi:type="ecore:EReference" name="alias" eType="#//Alias"
-        containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="exp" lowerBound="1" eType="#//Expression"
-        containment="true"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="FromItem" abstract="true" interface="true"/>
-  <eClassifiers xsi:type="ecore:EClass" name="Table" eSuperTypes="#//FromItem">
-    <eStructuralFeatures xsi:type="ecore:EReference" name="alias" eType="#//Alias"
-        containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EAttribute" name="name" lowerBound="1" eType="ecore:EDataType http://www.eclipse.org/emf/2002/Ecore#//EString"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="SubSelect" eSuperTypes="#//FromItem #//Expression">
-    <eStructuralFeatures xsi:type="ecore:EReference" name="alias" eType="#//Alias"
-        containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="selectBody" lowerBound="1"
-        eType="#//PlainSelect" containment="true"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="LongValue" eSuperTypes="#//Expression">
-    <eStructuralFeatures xsi:type="ecore:EAttribute" name="value" eType="ecore:EDataType http://www.eclipse.org/emf/2002/Ecore#//ELongObject"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="IsNullExpression" eSuperTypes="#//Expression">
-    <eStructuralFeatures xsi:type="ecore:EReference" name="exp" lowerBound="1" eType="#//Expression"
-        containment="true"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="CountAllFunction" eSuperTypes="#//Expression"/>
-  <eClassifiers xsi:type="ecore:EClass" name="Column" eSuperTypes="#//Expression">
-    <eStructuralFeatures xsi:type="ecore:EAttribute" name="name" eType="ecore:EDataType http://www.eclipse.org/emf/2002/Ecore#//EString"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="table" eType="#//Table"
-        containment="true"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="CaseExpression" eSuperTypes="#//Expression">
-    <eStructuralFeatures xsi:type="ecore:EReference" name="caseExp" eType="#//Expression"
-        containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="elseExp" eType="#//Expression"
-        containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="whenClauses" lowerBound="1"
-        upperBound="-1" eType="#//WhenClause" containment="true"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="StringValue" eSuperTypes="#//Expression">
-    <eStructuralFeatures xsi:type="ecore:EAttribute" name="value" eType="ecore:EDataType http://www.eclipse.org/emf/2002/Ecore#//EString"/>
-  </eClassifiers>
-  <eClassifiers xsi:type="ecore:EClass" name="WhenClause">
-    <eStructuralFeatures xsi:type="ecore:EReference" name="whenExp" lowerBound="1"
-        eType="#//Expression" containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="thenExp" lowerBound="1"
-        eType="#//Expression" containment="true"/>
-  </eClassifiers>
-</ecore:EPackage>
